Reorganize file structure to encapsulate internal code.
